chore 📌: Updates Flutter SDK version to 3.27.0
Updates the Flutter SDK version used by FVM to 3.27.0 in both the root and example directories. This ensures the project uses the latest features and improvements of the Flutter framework.

Removes the now redundant VS Code settings file from the project root, as the relevant settings are already defined in the example project's VS Code settings.

Additionally, the commit updates the .gitignore files to exclude the FVM version cache and Swift Package Manager files.

# Conflicts:
#	.vscode/settings.json